+++ src/estudante/entities/estudante.entity.ts
@@ -27,10 +27,10 @@ export class Estudante {
   cursoInteresse: string;
 
   @Column({
-  type: 'tinyint',
-  width: 1,
-  default: 1,
-})
+    type: 'smallint',
+    width: 1,
+    default: 1,
+  })
   ativo: boolean;
 
   @ManyToOne(() => Bolsa, (bolsa) => bolsa.estudante)
